refactor(db): extract SeaORM entities
