+++ crates/openwave-core/src/db/entities.rs
@@ -1,0 +1,219 @@
+pub mod document_generation {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "document_generation")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub document_id: Uuid,
+        pub content_revision: i64,
+        pub revision_token: Uuid,
+        pub tombstone: bool,
+        pub retirement_pending: bool,
+        pub retirement_content_revision: Option<i64>,
+        pub retirement_revision_token: Option<Uuid>,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod document {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "document")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub project_id: Option<Uuid>,
+        pub source_uri: Option<String>,
+        pub media_type: String,
+        pub title: Option<String>,
+        pub source_blob_id: Option<Uuid>,
+        pub source_sha256: Option<Vec<u8>>,
+        pub source_byte_len: Option<i64>,
+        #[sea_orm(column_type = "Text")]
+        pub canonical_text: String,
+        pub canonical_fingerprint: Option<String>,
+        pub source_regions: Json,
+        pub content_revision: i64,
+        pub revision_token: Uuid,
+        pub processing_status: String,
+        pub indexed_revision: Option<i64>,
+        pub index_fingerprint: Option<String>,
+        pub created_at: DateTimeUtc,
+        pub updated_at: DateTimeUtc,
+        pub indexed_at: Option<DateTimeUtc>,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod document_job {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "document_job")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub document_id: Uuid,
+        pub content_revision: i64,
+        pub revision_token: Uuid,
+        pub kind: String,
+        pub status: String,
+        pub pipeline_fingerprint: String,
+        pub attempt_count: i32,
+        pub max_attempts: i32,
+        pub available_at: DateTimeUtc,
+        pub lease_token: Option<Uuid>,
+        pub lease_expires_at: Option<DateTimeUtc>,
+        pub started_at: Option<DateTimeUtc>,
+        pub finished_at: Option<DateTimeUtc>,
+        pub last_error_code: Option<String>,
+        pub last_error_detail: Option<String>,
+        pub created_at: DateTimeUtc,
+        pub updated_at: DateTimeUtc,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod project {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "project")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub title: Option<String>,
+        pub workspace_dir: String,
+        pub created_at: DateTimeUtc,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod chat {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "chat")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub project_id: Option<Uuid>,
+        pub title: Option<String>,
+        pub model: Option<String>,
+        pub workspace_dir: String,
+        pub created_at: DateTimeUtc,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod message {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "message")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub chat_id: Uuid,
+        pub turn_id: Uuid,
+        pub role: String,
+        pub content: String,
+        pub created_at: DateTimeUtc,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod tool_call {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "tool_call")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub id: Uuid,
+        pub chat_id: Uuid,
+        pub turn_id: Uuid,
+        pub provider_id: String,
+        pub name: String,
+        #[sea_orm(column_type = "JsonBinary")]
+        pub arguments: Json,
+        pub result: Option<String>,
+        pub is_error: bool,
+        pub created_at: DateTimeUtc,
+        pub completed_at: Option<DateTimeUtc>,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod setting {
+    use sea_orm::entity::prelude::*;
+
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "setting")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub key: String,
+        // Matches the migration's `.json_binary()` (JSONB on Postgres).
+        #[sea_orm(column_type = "JsonBinary")]
+        pub value_json: Json,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
+
+pub mod event {
+    use sea_orm::entity::prelude::*;
+
+    // Composite primary key `(chat_id, seq)`: `seq` is monotonic *per
+    // chat*, and the pair both enforces uniqueness and indexes the
+    // "this chat's events after a cursor" replay query.
+    #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
+    #[sea_orm(table_name = "event")]
+    pub struct Model {
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub chat_id: Uuid,
+        #[sea_orm(primary_key, auto_increment = false)]
+        pub seq: i64,
+        #[sea_orm(column_type = "JsonBinary")]
+        pub payload: Json,
+        pub created_at: DateTimeUtc,
+    }
+
+    #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+    pub enum Relation {}
+
+    impl ActiveModelBehavior for ActiveModel {}
+}
